More intuitive arrow navigation.
Using the arrow keys to navigate elements, when editing the element value, will only work if the cursor is in the first or last position of the input element. 

Example: When the entire input value is selected, hitting the right arrow will drop the cursor at the end of the input as you would expect it to. When the cursor is at the end of the input and you hit the right arrow, it will then move to the next element.